+++ lib/cinder/library/filesystem.ex
@@ -10,4 +10,5 @@ defmodule Cinder.Library.Filesystem do
               {:ok, [{String.t(), non_neg_integer()}]} | {:error, term()}
   @callback mkdir_p(dir :: String.t()) :: :ok | {:error, term()}
   @callback ln(source :: String.t(), dest :: String.t()) :: :ok | {:error, term()}
+  @callback lstat(path :: String.t()) :: {:ok, File.Stat.t()} | {:error, term()}
 end